chore: move system monitor from pkg to internal
