+++ core/internal/monitor/buffer.go
