+++ core/internal/monitor/system_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/shirou/gopsutil/v4/disk"
 	"github.com/stretchr/testify/require"
-	"github.com/wandb/wandb/core/pkg/monitor"
+	"github.com/wandb/wandb/core/internal/monitor"
 	spb "github.com/wandb/wandb/core/pkg/service_go_proto"
 )
