+++ core/internal/monitor/tpuproto/generate.go
