+++ core/internal/monitor/tpuproto/generate.sh
@@ -1,5 +1,5 @@
 # Use import paths relative to the project root, which works better with IDEs.
-proto_path="${PWD%'core/pkg/monitor/tpuproto'}"
+proto_path="${PWD%'core/internal/monitor/tpuproto'}"
 protoc -I"$proto_path" \
     --go_out="$proto_path" \
     --go_opt=paths=source_relative \
