+++ core/pkg/server/server.go
@@ -12,9 +12,9 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/wandb/wandb/core/internal/monitor"
 	"github.com/wandb/wandb/core/internal/sentry_ext"
 	"github.com/wandb/wandb/core/internal/stream"
-	"github.com/wandb/wandb/core/pkg/monitor"
 	"github.com/wandb/wandb/core/pkg/server/listeners"
 )
